pyproject.toml
[tool.poetry]
name = "imgtopdfeasy"
version = "0.2.2"
description = "This is command line utility to convert images in a directory to  PDF file."
authors = ["Naveen <naveen@syrusdark.website>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/naveen521kk/img2pdf"

[tool.poetry.dependencies]
python = "^3.5"
Pillow = "^7.1.2"

[tool.poetry.plugins] # Optional super table
[tool.poetry.plugins."console_scripts"]
"img2pdf" = "imgtopdfeasy:main"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
